requirements.txt
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
jsonschema==4.20.0
pandas==2.1.4
numpy==1.24.3
pathlib2==2.3.7
typing-extensions==4.8.0 